sleeping_barber_pr_1

+++ tasks/mpi/gordeeva_t_sleeping_barber/func_tests/main.cpp
@@ -19,7 +19,7 @@ TEST(gordeeva_t_sleeping_barber_mpi, Test_Barber_0_Seats_Client) {
   gordeeva_t_sleeping_barber_mpi::TestMPITaskParallel testMpiTaskParallel(taskDataPar);
 
   if (world.rank() == 0) {
-    taskDataPar->inputs_count = {max_waiting_chairs_, barber_busy_};
+    taskDataPar->inputs_count = {max_waiting_chairs_, static_cast<const unsigned int>(barber_busy_)};
     taskDataPar->outputs.emplace_back(reinterpret_cast<uint8_t*>(global_res.data()));
     taskDataPar->outputs_count.emplace_back(global_res.size());
 
@@ -45,7 +45,7 @@ TEST(gordeeva_t_sleeping_barber_mpi, Test_Barber_3_Seats_Client) {
   gordeeva_t_sleeping_barber_mpi::TestMPITaskParallel testMpiTaskParallel(taskDataPar);
 
   if (world.rank() == 0) {
-    taskDataPar->inputs_count = {max_waiting_chairs_, barber_busy_};
+    taskDataPar->inputs_count = {max_waiting_chairs_, static_cast<const unsigned int>(barber_busy_)};
     taskDataPar->outputs.emplace_back(reinterpret_cast<uint8_t*>(global_res.data()));
     taskDataPar->outputs_count.emplace_back(global_res.size());
 
@@ -71,7 +71,7 @@ TEST(gordeeva_t_sleeping_barber_mpi, Test_Barber_15_Seats_Client) {
   gordeeva_t_sleeping_barber_mpi::TestMPITaskParallel testMpiTaskParallel(taskDataPar);
 
   if (world.rank() == 0) {
-    taskDataPar->inputs_count = {max_waiting_chairs_, barber_busy_};
+    taskDataPar->inputs_count = {max_waiting_chairs_, static_cast<const unsigned int>(barber_busy_)};
     taskDataPar->outputs.emplace_back(reinterpret_cast<uint8_t*>(global_res.data()));
     taskDataPar->outputs_count.emplace_back(global_res.size());
 
@@ -97,7 +97,7 @@ TEST(gordeeva_t_sleeping_barber_mpi, Test_Barber_50_Seats_Client) {
   gordeeva_t_sleeping_barber_mpi::TestMPITaskParallel testMpiTaskParallel(taskDataPar);
 
   if (world.rank() == 0) {
-    taskDataPar->inputs_count = {max_waiting_chairs_, barber_busy_};
+    taskDataPar->inputs_count = {max_waiting_chairs_, static_cast<const unsigned int>(barber_busy_)};
     taskDataPar->outputs.emplace_back(reinterpret_cast<uint8_t*>(global_res.data()));
     taskDataPar->outputs_count.emplace_back(global_res.size());
 
@@ -123,7 +123,7 @@ TEST(gordeeva_t_sleeping_barber_mpi, Test_Barber_100_Seats_Client) {
   gordeeva_t_sleeping_barber_mpi::TestMPITaskParallel testMpiTaskParallel(taskDataPar);
 
   if (world.rank() == 0) {
-    taskDataPar->inputs_count = {max_waiting_chairs_, barber_busy_};
+    taskDataPar->inputs_count = {max_waiting_chairs_, static_cast<const unsigned int>(barber_busy_)};
     taskDataPar->outputs.emplace_back(reinterpret_cast<uint8_t*>(global_res.data()));
     taskDataPar->outputs_count.emplace_back(global_res.size());
 
@@ -149,7 +149,7 @@ TEST(gordeeva_t_sleeping_barber_mpi, Test_Barber_500_Seats_Client) {
   gordeeva_t_sleeping_barber_mpi::TestMPITaskParallel testMpiTaskParallel(taskDataPar);
 
   if (world.rank() == 0) {
-    taskDataPar->inputs_count = {max_waiting_chairs_, barber_busy_};
+    taskDataPar->inputs_count = {max_waiting_chairs_, static_cast<const unsigned int>(barber_busy_)};
     taskDataPar->outputs.emplace_back(reinterpret_cast<uint8_t*>(global_res.data()));
     taskDataPar->outputs_count.emplace_back(global_res.size());
 
@@ -175,7 +175,7 @@ TEST(gordeeva_t_sleeping_barber_mpi, Test_Barber_Without_Seats_Client) {
   gordeeva_t_sleeping_barber_mpi::TestMPITaskParallel testMpiTaskParallel(taskDataPar);
 
   if (world.rank() == 0) {
-    taskDataPar->inputs_count = {barber_busy_};
+    taskDataPar->inputs_count = {static_cast<const unsigned int>(barber_busy_)};
     taskDataPar->outputs.emplace_back(reinterpret_cast<uint8_t*>(global_res.data()));
     taskDataPar->outputs_count.emplace_back(global_res.size());
