+++ tasks/mpi/gordeeva_t_sleeping_barber/perf_tests/main.cpp
@@ -17,7 +17,7 @@ TEST(gordeeva_t_sleeping_barber_mpi, test_pipeline_run) {
   std::shared_ptr<ppc::core::TaskData> taskDataPar = std::make_shared<ppc::core::TaskData>();
 
   if (world.rank() == 0) {
-    taskDataPar->inputs_count = {max_waiting_chairs, barber_busy_};
+    taskDataPar->inputs_count = {max_waiting_chairs, static_cast<const unsigned int>(barber_busy_)};
     taskDataPar->outputs.emplace_back(reinterpret_cast<uint8_t*>(global_res.data()));
     taskDataPar->outputs_count.emplace_back(global_res.size());
   }
@@ -55,7 +55,7 @@ TEST(gordeeva_t_sleeping_barber_mpi, test_task_run) {
   std::shared_ptr<ppc::core::TaskData> taskDataPar = std::make_shared<ppc::core::TaskData>();
   int num_clients = 10;
   if (world.rank() == 0) {
-    taskDataPar->inputs_count = {max_waiting_chairs, barber_busy_};
+    taskDataPar->inputs_count = {max_waiting_chairs, static_cast<const unsigned int>(barber_busy_)};
     taskDataPar->outputs.emplace_back(reinterpret_cast<uint8_t*>(global_res.data()));
     taskDataPar->outputs_count.emplace_back(global_res.size());
   }
